# Speedup cold runs of PSSA by using a runspace pool and parallelizing the slowest rule (AvoidAlias)