## PR Summary

We can decide to defer this performance improvement PR after the 1.18.0 release (or maybe merge only the runspace improvement?)

In total this can cut the time in half of analyzing a big script like e.g. [build.psm1](https://github.com/PowerShell/PowerShell/blob/master/build.psm1) the first time in a new shell.

- Cold runs (where the command info cache is not initialized) are sped up by 20% by re-using a runspace pool, thanks @daxian-dbw for the original idea. Creation and opening of it incurs a one-off initialization cost of around 50ms, which is negligible in most cases where the script is of a decent length (10-100 lines plus). A runspace pool of 6 is required for that due to 5 rules being able to make queries to it in parallel.
- The weakest link in the chain of PSSA runs is the slowest rule, I profiled rule execution and the `AvoidAlias` rule is the slowest rule. The slowness comes from a lot of calls to the command info cache, therefore this part is being parallelized and the runspace pool count is increased to 10 for that. This leads to a 30% improvement. The rule execution times before in this paraellization improvement changed the longest running rule from 7884 ms to 5164 (they are run in parallel, therefore the slowest rule is the weakest link):
```
Avoid Using Positional Parameters - Before: 3485ms - After: 4292ms
Should Process- Before: 4391ms - After: 5164ms
Use Cmdlet Correctly- Before: 4980ms - After: 4261 ms
Avoid Using Cmdlet Aliases or omitting the 'Get-' prefix. - Before: 7884ms - After: 3138ms
```
I did measure of course multiple times, the numbers are just an example of an average/normal run

## PR Checklist

- [x] [PR has a meaningful title](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---submission)
    - Use the present tense and imperative mood when describing your changes
- [x] [Summarized changes](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---submission)
- [x] [Change is not breaking](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#making-breaking-changes)
- [x] [Make sure all `.cs`, `.ps1` and `.psm1` files have the correct copyright header](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---submission)
- [ ] Make sure you've added a new test if existing tests do not effectively test the code changed and/or updated documentation
- [x] This PR is ready to merge and is not [Work in Progress](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---work-in-progress).
    - If the PR is work in progress, please add the prefix `WIP:` to the beginning of the title and remove the prefix when the PR is ready.